Move getPostTypeById to functions
