+++ lib/wordpress/_global/getPostTypeStaticProps.js
@@ -1,10 +1,10 @@
 import getFrontendPage from '@/functions/wordpress/postTypes/getFrontendPage'
 import getPostTypeArchive from '@/functions/wordpress/postTypes/getPostTypeArchive'
+import getPostTypeById from '@/functions/wordpress/postTypes/getPostTypeById'
 import {algoliaIndexName} from '@/lib/algolia/connector'
 import {addApolloState} from '@/lib/apolloConfig'
 import archiveQuerySeo from '@/lib/wordpress/_config/archiveQuerySeo'
 import frontendPageSeo from '@/lib/wordpress/_config/frontendPageSeo'
-import getPostTypeById from './getPostTypeById'
 import getPostTypeTaxonomyArchive from './getPostTypeTaxonomyArchive'
 import getSettingsCustomPage, {
   customPageQuerySeo
